First version of the website
Right now it is just a skeleton.
This is a simple template I made, with icons for the linkedin and github account.
Since this is a website for my portfolio, I also invested some time with animations and some cool effects with the css.
I just need to add more content.